refactor: detach avatar and bubble components
