+++ src/components/pages/ChatroomPage/Avatar.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+import { Avatar, AvatarProps, IMessage } from 'react-native-gifted-chat'
+
+export const CustomAvatar = (
+  props: Readonly<AvatarProps<IMessage>> &
+    Readonly<{
+      children?: React.ReactNode
+    }>
+) => {
+  return <Avatar {...props} />
+}
