+++ src/components/pages/index.ts
@@ -1,2 +1,3 @@
 export { default as LoadingPage } from './LoadingPage'
 export { default as SearchUserPage } from './SearchUserPage'
+export { default as ChatroomPage } from './ChatroomPage'
